+++ lib/src/custom_editor_provider.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:protobuf/protobuf.dart';
+import 'package:protobuf_message_editor/protobuf_message_editor.dart';
+
+abstract class CustomEditorProvider {
+  CustomFieldEditorBuilder? getCustomFieldBuilder(FieldIdentifier identifier) =>
+      null;
+
+  WidgetBuilder? getSubmessageEditorBuilder(
+    GeneratedMessage submessage,
+    GeneratedMessage? parentMessage,
+    FieldInfo? fieldInfo,
+  ) => null;
+
+  dynamic Function({
+    required FieldIdentifier identifier,
+    required FieldInfo fieldInfo,
+  })?
+  getRepeatedFieldAddBuilder() => null;
+}