+++ lib/src/proto_message_editor.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:protobuf/protobuf.dart';
+import 'package:protobuf_message_editor/protobuf_message_editor.dart';
 import 'package:protobuf_message_editor/src/custom_editor_registry.dart';
 import 'package:protobuf_message_editor/src/default_editors/default_editor_registry.dart';
 import 'package:protobuf_message_editor/src/field_editors/proto_field_editor.dart';
@@ -16,27 +17,33 @@ typedef SubmessageBuilder =
 
 class ProtoMessageEditor extends StatefulWidget {
   final GeneratedMessage message;
+  final GeneratedMessage? parentMessage;
+  final FieldInfo? fieldInfo;
 
   @Deprecated(
     "Users should wrap the ProtoMessageEditor in an ExpansionTile instead",
   )
   final String? expansionsTileTitle;
 
-  final CustomEditorRegistry? customEditorRegistry;
-  final SubmessageBuilder submessageBuilder;
+  final CustomEditorProvider? customEditorProvider;
+  final SubmessageBuilder? submessageBuilder;
   final VoidCallback? onRebuildRequested;
 
   const ProtoMessageEditor({
     super.key,
     required this.message,
+    this.parentMessage,
+    this.fieldInfo,
     this.expansionsTileTitle,
-    this.customEditorRegistry,
-    this.submessageBuilder = ProtoFieldEditor.defaultSubmessageBuilder,
+    this.customEditorProvider,
+    this.submessageBuilder,
     this.onRebuildRequested,
   });
 
   factory ProtoMessageEditor.withCustomEditors({
     required GeneratedMessage message,
+    GeneratedMessage? parentMessage,
+    FieldInfo? fieldInfo,
     String? expansionsTileTitle,
     Iterable<CustomFieldEditorBuilder> customFieldBuilders = const [],
     Iterable<CustomMessageEditorBuilder> customMessageEditors = const [],
@@ -46,7 +53,7 @@ class ProtoMessageEditor extends StatefulWidget {
     })?
     repeatedFieldAddBuilder,
   }) {
-    final customEditorRegistry = CustomEditorRegistry.fromIterable(
+    final customEditorProvider = CustomEditorRegistry.fromIterable(
       customFieldBuilders: customFieldBuilders,
       customMessageEditors: customMessageEditors,
       repeatedFieldAddBuilder: repeatedFieldAddBuilder,
@@ -54,8 +61,10 @@ class ProtoMessageEditor extends StatefulWidget {
 
     return ProtoMessageEditor(
       message: message,
+      parentMessage: parentMessage,
+      fieldInfo: fieldInfo,
       expansionsTileTitle: expansionsTileTitle,
-      customEditorRegistry: customEditorRegistry,
+      customEditorProvider: customEditorProvider,
     );
   }
 
@@ -64,42 +73,42 @@ class ProtoMessageEditor extends StatefulWidget {
 }
 
 class _ProtoMessageEditorState extends State<ProtoMessageEditor> {
-  late final CustomEditorRegistry customEditorRegistry;
-
-  @override
-  void initState() {
-    super.initState();
-    customEditorRegistry =
-        widget.customEditorRegistry ??
-        Provider.of<CustomEditorRegistry?>(context, listen: false) ??
-        const CustomEditorRegistry();
-  }
-
-  Widget _buildField(BuildContext context, FieldInfo fieldInfo) {
+  Widget _buildField(
+    BuildContext context,
+    FieldInfo fieldInfo,
+    CustomEditorProvider customEditorProvider,
+    SubmessageBuilder submessageBuilder,
+  ) {
     return ProtoFieldEditor(
       key: ValueKey((widget.message, fieldInfo.tagNumber)),
       message: widget.message,
       fieldInfo: fieldInfo,
-      repeatedFieldAddBuilder: customEditorRegistry.repeatedFieldAddBuilder,
-      submessageBuilder: widget.submessageBuilder,
+      repeatedFieldAddBuilder: customEditorProvider
+          .getRepeatedFieldAddBuilder(),
+      submessageBuilder: submessageBuilder,
       onRebuildRequested: widget.onRebuildRequested,
     );
   }
 
-  Widget _buildContent(BuildContext context) {
-    final customMessageEditor =
-        customEditorRegistry.getCustomMessageEditor(
-          widget.message.info_.qualifiedMessageName,
+  Widget _buildContent(
+    BuildContext context,
+    CustomEditorProvider customEditorProvider,
+    SubmessageBuilder submessageBuilder,
+  ) {
+    final customEditorBuilder =
+        customEditorProvider.getSubmessageEditorBuilder(
+          widget.message,
+          widget.parentMessage,
+          widget.fieldInfo,
         ) ??
-        defaultEditorRegistry.getCustomMessageEditor(
-          widget.message.info_.qualifiedMessageName,
+        defaultEditorRegistry.getSubmessageEditorBuilder(
+          widget.message,
+          widget.parentMessage,
+          widget.fieldInfo,
         );
-    if (customMessageEditor != null) {
-      return customMessageEditor.build(
-        context,
-        data: widget.message,
-        parentMessage: widget.message,
-      );
+
+    if (customEditorBuilder != null) {
+      return customEditorBuilder(context);
     }
 
     final fields = widget.message.info_.fieldInfo;
@@ -109,7 +118,7 @@ class _ProtoMessageEditorState extends State<ProtoMessageEditor> {
       );
 
       final customFieldBuilder =
-          customEditorRegistry.getCustomFieldBuilder(fieldIdentifier) ??
+          customEditorProvider.getCustomFieldBuilder(fieldIdentifier) ??
           defaultEditorRegistry.getCustomFieldBuilder(fieldIdentifier);
       if (customFieldBuilder != null) {
         return customFieldBuilder.build(context, parentMessage: widget.message);
@@ -117,7 +126,12 @@ class _ProtoMessageEditorState extends State<ProtoMessageEditor> {
 
       return Padding(
         padding: EdgeInsets.symmetric(vertical: 3),
-        child: _buildField(context, fieldInfo),
+        child: _buildField(
+          context,
+          fieldInfo,
+          customEditorProvider,
+          submessageBuilder,
+        ),
       );
     }).toList();
 
@@ -141,11 +155,30 @@ class _ProtoMessageEditorState extends State<ProtoMessageEditor> {
 
   @override
   Widget build(BuildContext context) {
-    final widgetCustom = widget.customEditorRegistry;
-    if (widgetCustom != null) {
-      return Provider.value(value: widgetCustom, child: _buildContent(context));
-    }
+    final effectiveCustomEditorProvider =
+        widget.customEditorProvider ??
+        Provider.of<CustomEditorProvider?>(context, listen: false) ??
+        const CustomEditorRegistry();
 
-    return _buildContent(context);
+    final effectiveSubmessageBuilder =
+        widget.submessageBuilder ??
+        Provider.of<SubmessageBuilder?>(context, listen: false) ??
+        ProtoFieldEditor.defaultSubmessageBuilder;
+
+    final content = _buildContent(
+      context,
+      effectiveCustomEditorProvider,
+      effectiveSubmessageBuilder,
+    );
+
+    return MultiProvider(
+      providers: [
+        Provider<CustomEditorProvider>.value(
+          value: effectiveCustomEditorProvider,
+        ),
+        Provider<SubmessageBuilder>.value(value: effectiveSubmessageBuilder),
+      ],
+      child: content,
+    );
   }
 }